ServiceVersion/requirements.txt
Django==2.2.9
xmltodict==0.12.0
requests==2.22.0
djangorestframework==3.11.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.8
more-itertools==8.1.0
packaging==20.0
pip==19.2.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
pytz==2019.3
setuptools==41.2.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.8